requirements.txt
google-genai
discord.py
python-dotenv
flask
pymongo
certifi
beautifulsoup4
requests
youtube-transcript-api
pypdf
python-docx
duckduckgo-search
dateparser
pytz
yfinance
elevenlabs